Add GSL diag tables to parm/ufs/fv3
 Update GSL template for hera

 Create EMC and GSL versions of files under workflow/rocoto
  - tasks.py, workflow_tasks.py
